Move things into i_slint_core::textlayout::sharedparley
